requirements.txt
scrython==1.7.1
discord.py==1.2.5
nest_asyncio==1.2.2